# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BIS API Platform
  version: 1.0.0
  summary: Unified payments, messaging, and provider orchestration API.
  description: |
    The BIS API Platform exposes a single, versioned surface for initiating
    payments, sending messages, managing providers, and receiving asynchronous
    event notifications (webhooks).

    All requests are routed through a central API gateway which handles
    authentication, rate limiting, request/correlation IDs, and dispatch to the
    appropriate provider adapter. Client applications never interact with
    providers directly; provider topology changes are invisible to API consumers.

    ## Base URL
    - Production: `https://api.company.com/v1`
    - Sandbox:   `https://sandbox.api.company.com/v1`

    ## Conventions
    - All request and response bodies are JSON (`Content-Type: application/json`).
    - All timestamps are RFC 3339 / ISO 8601 UTC strings.
    - All monetary amounts are expressed in the smallest currency unit
      (e.g. cents) unless explicitly noted, and are always accompanied by a
      `currency` ISO 4217 code.
    - Identifiers are prefixed strings (e.g. `pay_...`, `msg_...`, `ref_...`,
      `conv_...`, `prov_...`).

    ## Cross-cutting headers
    Every response includes the following headers (see `parameters` / `responses`
    sections for per-operation detail):

    | Header             | Description                                                          |
...

...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /providers/{id}:
    parameters:
      - $ref: '#/components/parameters/ProviderId'
    get:
      tags: [Providers]
      operationId: getProvider
      summary: Retrieve a provider
      description: Retrieve a single provider's public configuration and live health.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The provider.
          headers:
            X-Request-Id: { $ref: '#/components/parameters/RequestIdHeader' }
            X-Correlation-Id: { $ref: '#/components/parameters/CorrelationIdHeader' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Provider' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /webhooks/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        description: |
          Provider identifier whose webhook is being delivered
          (e.g. `stripe`, `flutterwave`, `pawapay`, `paychangu`, `infobip`,
          `signalhouse`).
        schema: { type: string }
    post:
      tags: [Webhooks]
      operationId: receiveProviderWebhook
      summary: Receive an inbound provider webhook
      description: |
        Provider-agnostic inbound webhook receiver. This endpoint is called by
        upstream providers (not by API consumers).

        The raw request body is captured before parsing for HMAC/JWT signature
        verification. Each provider uses a distinct signature scheme — see the
        `WEBHOOK_ARCHITECTURE` doc. Verification failures return `401`.

        **Idempotency:** each event is deduplicated by
        `(provider, provider_event_id)`. Duplicate deliveries return `200`
        without re-processing.
      security: []
      parameters:
        - name: stripe-signature
          in: header
          description: Stripe `Stripe-Signature` header (HMAC).
          schema: { type: string }
        - name: verif-hash
          in: header
          description: Flutterwave `verif-hash` header.
          schema: { type: string }
        - name: x-paychangu-signature
          in: header
          description: PayChangu `X-PayChangu-Signature` header (`sha256=...`).
          schema: { type: string }
        - name: Authorization
          in: header
          description: PawaPay JWT bearer token.
          schema: { type: string }
      requestBody:
        required: true
        description: Raw provider payload. Consumed as `application/json` after signature verification.
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: |
            Accepted. Returned immediately after persistence; async processing
            (status update, event emission) happens after the response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean, example: true }
        '401':
          description: Signature verification failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Unknown provider.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: |
        API key issued per application. Send as a Bearer token:

        ```
        Authorization: Bearer sk_live_••••••••••1234
        ```

        Use `sk_test_...` keys against the sandbox. Keys are scoped to a single
        `app_id`; the gateway validates the key against the `api_keys` table and
        rejects requests whose `app_id` does not match the key's scope.
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      description: Unique UUID for this HTTP request.
      schema: { type: string, format: uuid }
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      description: |
        Correlation ID tracing a logical operation across services and async
        webhooks. Echoed back from any webhook events derived from this request.
      schema: { type: string }
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Optional correlation ID. If omitted, the gateway generates one.
      schema: { type: string }
...
    ProviderId:
      name: id
      in: path
      required: true
      description: Provider identifier (e.g. `stripe`, `pawapay`).
      schema: { type: string }
  responses:
...
    Unauthorized:
      description: Authentication failed or key not authorized for `app_id`.
      headers:
        X-Request-Id: { $ref: '#/components/parameters/RequestIdHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: authentication_failed
              message: "Invalid or missing API key"
              request_id: "req_9z8y7x6w"
    NotFound:
      description: Resource not found.
      headers:
        X-Request-Id: { $ref: '#/components/parameters/RequestIdHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: not_found
              message: "Payment pay_8f2c1a9b not found"
              request_id: "req_aa11bb22"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              enum:
                - invalid_request
                - authentication_failed
                - permission_denied
                - not_found
                - invalid_operation
                - idempotency_conflict
                - rate_limited
                - provider_error
                - internal_error
            message:
              type: string
              description: Human-readable description.
            request_id:
              type: string
              description: Echo of `X-Request-Id` for support correlation.
            correlation_id:
              type: string
              description: Echo of `X-Correlation-Id` if supplied.
            details:
              type: array
              description: Field-level validation errors.
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
    ProviderCategory:
      type: string
      enum: [payment, messaging, other]
    ProviderStatus:
      type: string
      enum: [online, offline, maintenance]
    ProviderHealthStatus:
      type: string
      enum: [healthy, degraded, down, unknown]
...
    Provider:
      type: object
      required: [id, name, category, status, capabilities]
      properties:
        id: { type: string, example: pawapay }
        name: { type: string, example: "PawaPay" }
        category:
          $ref: '#/components/schemas/ProviderCategory'
        status:
          $ref: '#/components/schemas/ProviderStatus'
        environment:
          type: string
          enum: [test, live]
        countries:
          type: array
          description: Supported ISO country codes, or `['*']` for global.
          items: { type: string }
          example: ["MW", "ZM", "TZ"]
        currencies:
          type: array
          items: { type: string }
          example: ["MWK", "ZMW", "TZS"]
        capabilities:
          type: array
          items: { type: string }
          example: ["mobile_money"]
        weight:
          type: number
          description: Routing priority (0-100).
        health:
          $ref: '#/components/schemas/ProviderHealthStatus'
        last_successful_request:
          type: string
          nullable: true
          format: date-time
        error_rate:
          type: number
          description: Rolling error rate percentage (0-100).
        latency_estimate_ms:
          type: object
          properties:
            min: { type: number }
            max: { type: number }
...